Agregado bootstrap3 con ayuda de una pagina guardada en evernote, estoy mejorando el disenio de mi formulario

+++ db/migrate/20131129031543_create_archivos.rb
@@ -3,9 +3,9 @@ class CreateArchivos < ActiveRecord::Migration
     create_table :archivos do |t|
       t.string :nombre
 
-      t.integer :color_r
-      t.integer :color_g
-      t.integer :color_b
+      t.decimal :color_r
+      t.decimal :color_g
+      t.decimal :color_b
 
       t.integer :num_lados
       t.integer :size_lados
